% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton.R
\name{newton}
\alias{newton}
\title{Approximate solution of f(x) = 0 by Newton's method.}
\usage{
newton(f, guess, iter = 1000)
}
\arguments{
\item{f}{Function for which we are searching for a solution f(x) = 0.}

\item{guess}{Initial guess for a solution f(x) = 0.}

\item{iter}{Iterations.}
}
\value{
The estimated root of the equations.
}
\description{
Pick a distribution based on a list of constraints.
}
\keyword{internal}
