% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\docType{class}
\name{Normal-class}
\alias{Normal-class}
\alias{Normal}
\title{Constraint-based normal distribution}
\description{
A distribution class pairing a description in model syntax and a ggplot2 function.
}

\keyword{internal}
