% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_parameters.R
\name{pick_parameters}
\alias{pick_parameters}
\title{Pick parameters}
\usage{
pick_parameters(distribution, constraints, guess, epsilon, max_iter)
}
\arguments{
\item{constraints}{List of constraints.}
}
\value{
A distribution object.
}
\description{
Pick paramters for a given distribution based on a list of constraints.
}
\keyword{internal}
