% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\docType{class}
\name{Poisson-class}
\alias{Poisson-class}
\alias{Poisson}
\title{Constraint-based Poisson distribution}
\arguments{
\item{constraints}{List of constraints.}
}
\value{
Parameters for a Poisson distribution.
}
\description{
Find suitable parameters for a Poisson distributions given constraints.
}

\keyword{internal}
