% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{remove_x_axis}
\alias{remove_x_axis}
\title{Remove x axis
ggplot2 theme settings to completely remove the x axis.}
\usage{
remove_x_axis()
}
\value{
ggplot2 theme.
}
\description{
Remove x axis
ggplot2 theme settings to completely remove the x axis.
}
\keyword{internal}
