% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{remove_y_axis}
\alias{remove_y_axis}
\title{Remove y axis
ggplot2 theme settings to completely remove the y axis.}
\usage{
remove_y_axis()
}
\value{
ggplot2 theme.
}
\description{
Remove y axis
ggplot2 theme settings to completely remove the y axis.
}
\keyword{internal}
