% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server logic for priordist
Define the priordist server logic.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}
}
\value{
Shiny server logic.
}
\description{
Server logic for priordist
Define the priordist server logic.
}
\keyword{internal}
