% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{User interface for priordist}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
ui
}
\value{
A Shiny user-interface definition.
}
\description{
Define the priordist user interface.
}
\keyword{internal}
