% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson.R
\docType{class}
\name{Poisson-class}
\alias{Poisson-class}
\alias{Poisson}
\title{Constraint-based Poisson distribution}
\description{
A distribution class pairing a description in model syntax and a ggplot2 function.
}

\keyword{internal}
