% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server logic for priordist}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Provided by shiny.}

\item{output}{Provided by shiny.}

\item{session}{Provided by shiny.}
}
\value{
Shiny server logic.
}
\description{
Define the priordist server logic.
}
\keyword{internal}
